+++ juriscraper/opinions/united_states/federal_appellate/ca4.py
@@ -26,7 +26,6 @@ class Site(OpinionSiteLinear):
         self.parameters = {}
         self.update_parameters()
         self.make_backscrape_iterable(kwargs)
-        self.should_have_results = True
 
     def make_backscrape_iterable(self, kwargs: dict[str, str]) -> None:
         """Make back scrape iterable